adds umd to wrap code on build
