+++ package.json
@@ -4,7 +4,8 @@
   "description": "Easily convert an array to an object.",
   "main": "toObject.js",
   "scripts": {
-    "test": "jasmine"
+    "test": "jasmine",
+    "build": "npm test && node_modules/umd/bin/cli.js toObject src/toObject.js dist/toObject.js -c"
   },
   "nyc": {
     "exclude": [
@@ -24,6 +25,7 @@
   },
   "homepage": "https://github.com/theZieger/toObject#readme",
   "devDependencies": {
-    "jasmine": "^2.6.0"
+    "jasmine": "^2.6.0",
+    "umd": "^3.0.1"
   }
 }
